Fixed bug where clearing qos preemption is interpreted as qos id 0
Clearing out the list of qos' to preempt with preempt='' would result in
attempting to preempt the qos with an id of 0.  This occured since a
NULL string was being processed as the desired qos. This change instead
looks for the null character and clears the entire preempt bitstr
instead of clearing it then interpreting the NULL string.

Ticket 20737