Docs - Clarify semantics of DNS SRV TTL with configless
This patch adds a line in the NOTES section of configless_slurm.html to
clarify the fact that config fetch is a one-time op and no automatic
refresh is performed.

The fact that the slurmctld was discovered through DNS does not imply
an automatic refresh when the record's TTL expires. This has caused
some confusion, which this change addresses.

Ticket 20462