Make _top_priority() consider FLEX and ANY_NODES reservations
This fixes a long standing issue which it was assumed that if one job
had a reservation and the other did not then they would not overlap in
nodes at all. This, however, is not true for FLEX and ANY_NODES
reservations. The previous logic made it possible for a newly submitted job
without a reservation to delay a old pending job with a FLEX or ANY_NODES
reservation.

Ticket 20893