task/cgroup - Implement OOMKillStep logic
This commits adds the check of the cgroup feature memory.oom.group
together with the OOMKillStep implementation. When the cgroup feature
is available (the file memory.oom.group exists in the cgroup), and
the config is set, then at step_creation time the value 1 is set to
memory.oom.group at the step user level (where all step user processes
exist). That makes that in case of one of the tasks OOM's, all the
processes in the cgroup are terminated, thus making that the stepd
process reaches it's end of life. In that moment the check for OOM is
done, and if it is true a termination signal is sent to the step. That
makes that if this step is multinode, the processes in the other nodes
are terminated too.

Issue #50069