cgroup/v2 - Add new case in cgroup_p_has_feature CG_FALSE_ROOT
This feature check checks whether the mounted root cgroup is a real
root cgroup or just a domain from another cgroup. This is useful in the
situation that slurm has been started in a container with a namespaced
cgroup and mounted cgroup, and the mountpoint is a subdirectory in the
host.

A subdirectory of the root cgroup does not behave equal and has some
restrictions about what pids it can allocate.

Issue #50068