accounting_storage/mysql - fix delete assocs with empty partitions
Previously, deleting an assoc given with an empty partition would delete all
associations, not just the one with no partition. Example: If user1 has
acct1 associations with and without partitions then this would remove all
of them (not just the one with no partition):

  $ sacctmgr delete user user1 account=acct1 partition=\"\"

Make it so that only the assoc with no partition is deleted when an empty
partition is specified.

The previous behavior can still be achieved by not specifing any partition.

Ticket 19272.