cgroup/v2 - Add get_scope_path function
Retrieve the scope path from the cgroup plugin.

In the next commits this scope path will be retrieved from the slurmd
cgroup plugin and passed into the slurmstepd using the socket between them.

By having this information in the stepd we can move it into the appropiate
cgroup, plus in some situations (containers) we could not be able to infer
the correct path.

Issue #50068