cgroup/v2 - Split init into init and cgroup_p_setup_scope
This split allows to pass the scope_path parameter on plugin init.

For the slurmd init this will be NULL as is in slurmd init where we
create it. In next commits the scope_path will be used to determine
the cgroup where the slurmstepd process should be located.

Issue #50068