salloc - Remove --get-user-env option
The combination of commits 47d8784b and 3c1aac3f4 rendered --get-user-env
for salloc unusable. Please note that since --uid does not exist anymore,
--get-user-env will not be able to target other UIDs as it is.

Salloc used with --get-user-env retrieved the environment locally and
not in the allocated node like sbatch does, so this combined with the
reasons above, leaves this option with little reason to exist.

We are also removing related functions that would become unused due
to these changes.

Ticket 20987