cgroup/v2 - Move processes away from a false root cgroup
In the case that slurmd detects a false root cgroup (as in certain
container setup), move all the processes to a new cgroup /system.

slurmd will then be able to create the scope hierarchy to allocate the
stepd scope.

This only applies with the IgnoreSystemd option, as we do not want to
move processes around if systemd is in control.

Issue #50068