Fix slurmd invalid registration with e-cores and CoreSpec
On a system with p-cores and e-cores, the _remove_ecores function does
not work correctly as hwloc apparently does not return proper values
while identifying the different kinds of cpus.

If we are in the process of daemonizing slurmd, the original process will
create the xml from the whole node, then will constrain itself into a
cpuset by initializing cgroups, and then it will daemonize.
By daemonizing it will fork and start a new slurmd that will repeat the
same process, but it will be constrained by the cgroup modified by the
grandparent from the beginning. This will cause hwloc to generate an
incorrect xml file.

The consequence of this problem caused invalid registrations of the node.

The fix here avoids regenerating the hwloc xml file when we are in the
process of daemonization.

Ticket 20869