seff - use TRESUsageInTot instead of TRESUsageInMax * ntasks
TresUsageInTot in the database is the sum of all the memory peaks of all
the tasks of the step. This value is useful as what is the maximum memory
consumption to expect.

TresUsageInMax is the maximum peak seen for any task.

Before, the maximum memory consumption to expect was infered by multiplying
TresUsageInMax * ntasks, which was an over-estimation.

This fix replaces the calculation by TresUsageInTot, which is a better
approximation.

Ticket 21127.