Refactor xgetnameinfo() to improve concurrency.
Attempt to find a cached record under a read lock, thus allowing multiple
threads to search concurrently.

On a miss, move the getnameinfo() outside of any locks. This allows
threads to search in parallel, and avoids blocking the cache for other
threads in case this call takes time to complete.

Once completed, acquire the write lock to insert or update the record.