Preempt jobs for license only when there is lack of that license
In _overlap_and_running_internal the job that used the same license as
the one requested was always considered as overlap. It makes sense when
we want to reserve licenses or when we create preemptee_candidates, but
if we are actually preempting jobs we should do that only if there are
not enough licenses of that type to run the job.

Ticket: 21200