Fix validation of expression inside parentheses for extra constraints
For an extra constraints job request, fix validation that the expression
inside of parentheses ends in a leaf. The previous logic only checked if
there were any children (leaves). This new logic still catches that case,
and also catches the case that the expression ends in an operator instead of
ending in a leaf.

Ticket 18133
Issue #50136