sh5util - Update sh5util to reject het/array component in arguments
The h5 files that acct_gather produces do not have het/array index appended
to them, so there is no simple way for sh5util to know which file
corresponds to a particular component of a het/array job.

Instead, the user should determine what the raw JobId of the het/array
component is, and use that as input to sh5util.

For example, with an array job of 2 components:

$ sacct -X -j 27 --format JobId,JobIdRaw
JobID        JobIDRaw
------------ ------------
27_1         28
27_2         27

$ sh5util -j 28 # works for 27_1
$ sh5util -j 27 # works for 27_2

Changelog: Alter sh5util -j to not allow array or het job ids.
Ticket: 17714