Fix creating step on lower version stepmgr
init_header() will set the msg's protocol version if it's not set to
the global working_cluster_rec's protocol_version. However, in
slurm_buffers_pack_msg(), init_header() is called after the body is
packed with pack_msg(). The protocol_version needs to be set earlier so
that the body can be packed properly.