Check RPC version in unpack_header().
The minimum RPC version check was already enforced here.
Add in a check for RPCs that are too new, which is currently handled
by check_header_version() instead.

Still need to return SLURM_ERROR here, as the calling paths expect it.

Move the forward_init() after the check so we can return a different
error directly here in the future without needing destroy_forward().