Fix slurmctld with accounting_storage/none trying to connect to slurmdbd
If errno is set to a non-zero value in some previous initialization code
and accounting_storage/none is used, then the non-zero value for errno will
be passed to assoc_mgr_init(). This will be considered an error, and
slurmctld will try to reconnect to the slurmdbd, even though the slurmdbd is
not being used.

Issue #25