Reject jobs using --signal=R: if Preemption=OFF on system
Jobs requesting --signal=R: need to be preemptable since they can use
resouces in reservations not belonging to them. If preemption is off for
the whole system then preemption will not work, so deny --signal=R jobs in
that case.

Changelog: Deny jobs using the "R:" option of --signal if PreemptMode=OFF
 globally.
Ticket: 20928