Docs - Preemption required and force requeue or cancel for --signal=R
If --signal=R:<sig_num>[@sig_time] is used, the global PreemptMode can not
be OFF. The --signal=R: option relies on preemption to not violate
MaxStartDelay on reservations. The preemption mode of this job will be
requeue if allowed else cancel. Partition preempt modes suspend and off are
ignored.

Ticket: 20928