Require jobs using --signal=R get preempted by requeue or cancel only
Jobs using --signal=R can run on reservations not belonging to them.
If another job wants to use the reservation then the resources need to be
returned. PreemptMode suspend and off are not compatible with that. If the
partitions preemption mode is not requeue or cancel force cancel.

Changelog: Force jobs using the "R:" option of --signal to be preemptable
 by requeue or cancel only. If PreemptMode on the partition or QOS is off
 or suspend, the job will default to using PreemptMode=cancel.
Ticket: 20928