Save pending steps to job_state.
Otherwise, on a reconfigure/restart, slurmd won't automatically
schedule and launch these. The srun process will eventually re-contact
the slurmctld, but there's a decent delay here, and the pending steps
may not be serviced in the expected order.

This also fixes a crash unpacking the state caused by changes in
f0921dd7f1a. The list_count() included pending steps, but they weren't
packed in the buffer, and the unpack side would attempt to unpack further
job records as if they were step records, leading to a fatal().

Co-authored-by: Tim McMullan <mcmullan@schedmd.com>