Simplify logic around invalid exc_nodes parsing.
With the best_effort argument set to true, node_name2bitmap() will never
return an error. This behavior was established way back in 7ef7afe6ed9 -
if a node doesn't exist in the config, Slurm is pretty unlikely to schedule
on it, so simply ignore those invalid entries.

Change the best_effort argument to false here - this does have the side-effect
of ignoring the alias list, but using the node alias in the --exclude list
was never documented and is inconsistent with the --nodelist handling.