Avoid slurmd crashing after failing to unpack SlurmdSpoolDir/cred_state
If a newer version of slurmd has been started and upgraded the file formats
in SlurmdSpoolDir, trying to start an older version of slurmd can result in
an unpack failure. The cred_job_list and/or cred_state_list would then be
NULL. Calling list_delete_all() on a NULL list crashes. Instead of crashing,
log a warning and return.

Issue 50154, Ticket 21432