slurmstepd - Fix follow-up open() after last iteration mkdirpath().
This permits a follow-up open() attempt after mkdirpath() in the case of
mkdirpath() success in the last iteration before the maximum is reached.

Also, it doesn't make sense to count ENOENT + mkdirpath() in a first time as
a real error since automatic recursive directory creation is an expected and
supported functionality.

Ticket 18919

Co-authored-by: Carlos Tripiana Montes <tripiana@schedmd.com>