Fix backup controller crash after reconfiguring.
If the backup controller is in control and scontrol reconfigure is run,
after resuming the primary controller it was possible that the backup
would not honor the request to relinquish control.

In most cases this would result in the backup controller crashing, but
in a quiet enough system both controllers could continue to operate as
the primary.

Changelog: slurmctld - Fix crash and possible split brain issue if the
 backup controller handles an scontrol reconfigure while in control
 before the primary resumes operation.
Ticket: 21532
Cherry-picked: f37371bac8