acct_gather_energy/ipmi,xcc - Add comment about pthread_cancel
In the IPMI plugin, we on purpose call pthread_cancel just in case the
driver is stuck at some point. This is dangerous but is a best effort.
In the worst case we could end up with a segfault or a deadlock because
of glibc keeping locks or similar. Remind this in a comment.

Issue: 50172
Ticket: 6102