Unset default PTHREAD_CANCEL_ASYNCHRONUS in persist_conn.c
This was set in slurm_persist_conn_recv_server_init called by the
controller at main, and by rpc_mgr. This was doing all threads to be
asynchronously cancelable. We do not use any pthread_cancel in our code
which is relevant for this case and it seems this should not be
the default, as it is not always safe to pthread_cancel a thread at a
random point.

Issue: 50172
Ticket: 6102