cgroup/v2 - Sigkill the user processes step cgroup atomically
Where available, kill the entire cgroup by writting to cgroup.kill
instead of getting all the pids and killing them one by one. The previous
mode of operation, which is still done when cgroup.kill is not available
or when using cgroup/v1, was racy as a pid could show up in the cgroup
after we got the list. Moreover it was less performant.

This affects ../step/user cgroup which is where the user processes reside,
not slurmstepd processes.

Ticket: 14274
Issue: 50173