fatal() if slurmctld is not executed as SlurmUser when --systemd is used
become_slurm_user() is not called when launching with systemd.  This
lead to a situation where the slurmctld could be launched as a user
other than SlurmUser and cause various issues, including possibly
changing permissions on the state save files.

Prevent this situation by fataling the slurmctld if it is started
with --systemd and executed as a user other then SlurmUser.

Changelog: slurmctld - fatal() if the slurmctld is not running as
 SlurmUser when started with systemd.
Ticket: 21553