Fix stepmgr not getting dynamic node addrs from the controller
The controller doesn't pack job_record_t's node_addrs, so the stepmgr
wasn't getting them and passing them to the steps. When the steps
completed and tried communicating back to the dynamic stepmgr, it failed to
find its node addr. The job's node_addr's are passed in the cred so we
can just get them from there.

Changelog: Fix stepmgr not getting dynamic node addrs from the controller
Ticket: 21535