Remove implicit format_stderr in thread_id LogFormat
There was a block of code that implicitly set format_stderr when setting
thread_id. This caused thread_id to always format the stderr when one
could prefer to only format the log files and not the command output.

This was kept for backward compatibility reasons as in versions previous to
23.11 there was no way to format the stderr but to set thread_id.

This is not needed anymore and if thread_id will behave now like all the
other LogTimeFormat options.

Changelog: Remove implicit format_stderr in thread_id, so
 LogFormat=thread_id does not print to stderr anymore.
Issue: 50180
Co-authored-by: Marcin Stolarek <cinek@schedmd.com>