Fix not removing boot^ state on unexpected reboots
When a node is planned for reboot asap and suffers an unexpected reboot,
when it returns to service the controller reboots the node again and then
it never removes the boot^ flag, even if it comes back correctly.

Remove the flag and Reason when an unexpected reboot is detected and there
was a reboot planned in the node.

The node will be set to down.

Changelog: Remove the boot^ state on unexpected node reboot after
 return to service.
Ticket: 18087
Cherry-picked: 8f50a09f8b