Fix not removing a node from rs_node_bitmap when it is being rebooted
This caused jobs to be scheduled and started R while the node was
rebooting. This happened when the flag nextstate=resume was used.

The rs_node_bitmap was or'ed with avail_node_bitmap in _schedule()
allowing the main scheduler to schedule on the resuming nodes. We need
to not schedule on the resuming nodes and just adjust the message.

This bug was introduced in commit da827d45493.

Changelog: Do not allow new jobs to start on a node which is being rebooted
 with the flag nextstate=resume.
Ticket: 18087
Cherry-picked: f9f1b4f309