slurmdbd - Remove races while terminating rollup handler thread
Checking running_rollup variable could cause a race as it was not protected
by a mutex. Use of pthread_mutex_trylock will acquire the lock if it is not
acquired or return EBUSY if the rollup is being executed, which removes any
possible race.

This also restores behavior lost in commit c6d41c1086 where the backup DBD was
forced to block until a rollup was finished. The rollup is done on a
transaction so throwing it away and letting the primary do it is preferred.

Ticket: 21560
Issue: 50178