Docs - data_parser/v0.0.42: Note that argv[0] of job desc is overridden
The slurmstepd always overrides argv[0] with the path to the newly created
script file on the slurmd node for batch scripts. Since user might be
relying on the behavior of the slurmrestd jobs seemly throwing away
argv[0], this only makes note of the behavior in the documentation so that
we don't break users' scripts.

Ticket: 21597
Cherry-picked: 43cabf6fc1