task/cgroup - cpuset now works at task level
This is analogous to commit 18e5a51b but for cpuset cgroup.

There was a race when jobacctgather/cgroup moved a pid to task_x directory
in the cgroup controller. Then task/cgroup was removing it and putting
again the pid in the step_0 superior directory.

Do here what we already do for task/cgroup devices and memory and work at
task level.

The conflicting controllers in v1 were between task/cgroup and
jobacctgather/cgroup was only memory and devices, since the cpu was
gathered from cpuacct and task worked with cpuset, but in v2 this is
not valid anymore because of the unified hierarchy.

Ticket: 21263
Cherry-picked: 9f1cd828ff