Add changelog for the last 6 commits
Remove critical pthread_cancel in slurmdbd.

Now only 5 pthread_cancel calls remain in the slurm code.

Three of them are still a concern but should not be of great impact.

The first two are in ipmi and xcc, but we have no option other than
cancelling if the driver is stuck at some point and we want to stop the
thread.

In the case of the rollup thread in slurmdbd backup, we want it
to terminate immediately and not let it run while the primary slurmdbd might
be running also a rollup. We're anyway restarting the backup dbd after it.

Another one is ok as commit delay thread in slurmdbd is
cancelled only in a safe point.

The only remaining one is in acct_gather_energy_gpu.c but it not needed at all
but will be handled in a different MR.

Issue: 50172
Ticket: 6102