Nodes that finish jobs are given to replace reservations
It was possible for nodes to be assigned to high priority jobs
before a reservation with the replace flag could claim nodes.
Reservations are considered to be the highest priority job, so once
a job has finished executing the controller will check if the node
can be set to a replace reservation.

Ticket: 21596