Fix job requeuing when all nodes in a job fail simultaneously
When a job taking 2 or more nodes had all of its nodes fail, and no
EpilogSlurmctld was configured, job requeuing was not correctly processed
as batch_requeue_fini was not called. This resulted in the following issues:

- Requeued job was not assigned a new DB index.
- Job steps of new jobs were not being reset to 0.

This left incorrect entries in the accounting database for the requeued job.
Added a batch_requeue_fini call to fix that.

Ticket: 20177