slurmdbd - Remove non-async-cancel safe call during rollup when shutdown
When shutting down threads in slurmdbd we were calling pthread_cancel
for the rollup thread. As the thread accepted async cancel calls, it could
be cancelled at any point, even when glibc had locked some locks. This
could cause deadlocks, segfaults or memleaks.

Move to a safe pthread_cond_signal approach for this thread for the
primary controller.

For the backup slurmdbd, we keep forcibly killing the thread.

Issue: 50172
Ticket: 6102