acct_gather_energy/gpu - Remove all the thread launching logic
This code creates a thread that launches the thread_gpu_run. All of this
is not needed in the gpu plugin and the _thread_gpu_run for gathering
energy can be directly launched.

This was a bad copy-paste from ipmi plugins where the launcher needs to
wait in case the ipmi is stuck, but in the gpu plugin there is nothing
that can get stuck so it is not needed.

Issue: 50208