data_parser/v0.0.43 - Mark instances as removed in /config
Before OPENAPI_SLURMDBD_CONFIG_RESP, which is used for /config endpoint,
had INSTANCE_LIST as one of its parsers. However, slurmrestd doesn't
retrieve the instances from the database during a /config query.
Additionally, instances should not be a part of /config output. This marks
INSTANCE_LIST as deprecated in 25.05.

Changelog: data_parser/v0.0.43 - Deprecating the 'instances' field of the
 GET /slurmdb/v0.0.43/config endpoint. This field was never populated.
Ticket: 21667