Docs - Improve --json --yaml man page descriptions
Add "most" to the filter descriptions for squeue, sinfo, and scontrol
argument `--json` and `--yaml` for support of filtering. While the queries
are done almost entirely by the CLI command, the final stage filtering
that is done to the text output may not always be applied to the JSON or
YAML outputs.

Ticket: 18361