Fix segfault when testing jobs against nodes with invalid gres
Last job test - in case all previous tests failed - will check if the
job fits in ANY of the nodes in the cluster (including INVALID nodes).

This prevents a possible segfault when checking for gres bitmap in nodes
that did not configure gres correctly and were marked as INVALID.

Ticket: 21556
Changelog: Fix segfault when testing jobs against nodes with invalid gres.