Remove mode parameter from env_array_user_default
This function is called just once in all the codebase, which hardcodes
the mode parameter to 0. This effectively leaves the "mode" parameter
without any actual use, so we are removing it from the function.

Consequently, since we know that the mode will always be the same, we
have removed its related logic from env_array_user_default and any
downward functions that are only called there, like _child_fn.

Ticket: 5037