select/cons_tres - Fix crash and issues evaluating jobs
Suspended jobs do not get removed from node usage so, it you cancel it after
that, there's a pointer there to a finished job.

This causes two issues:

1. Can prevent from running the evaluated job.
2. If the deleted job is purged, any attempts to read its contents will lead
to bad data and potential crash.

In the related ticket, _is_job_sharing was segfaulting.

Changelog: Fix crash and issues evaluating job's suitability for running in
 nodes with already suspended job(s) there.
Ticket: 21767