slurmctld - fatal() on startup if a core state file doesn't exist
If the clustername file exists but we can't open one of these state
files, it likely means something has gone wrong in the filesystem so
fatal() instead of continuing.

This is only done for state files that are always expected to be
present.

Changelog: slurmctld - fatal() on startup if certain core state files do
 not exist and it is not the first start of the cluster.
Issue: 50162