job_test_resv() - Limit available to those in the job's partition
For FLEX or ANY_NODES reservations, job_test_resv() can set *node_bitmap to
all nodes in the cluster (the result of node_conf_get_active_bitmap()). This
means that job_test_resv() can return with nodes set in *node_bitmap that
are not in the current partition for the job. This can mess up later testing
logic that assumes that all nodes in an available node bitmap are part of
the current partition.

For example, assume there are two reservations: one is a current
OVERLAP,FLEX reservation; the other is a future MAINT reservation that
entirely overlaps the current reservation. In _build_node_list(),
&usable_node_mask is passed to job_test_resv() as the nodelist parameter. If
the job's time limit intersects the start time of the future maintenance
reservation, job_test_resv() removes all overlapping nodes between the two
reservations. This means that usable_node_mask can have one or more nodes
set that are not in the current partition, but zero nodes set that are in
the current partition.

In this situation, the job fails the following check that would return with
ESLURM_RESERVATION_BUSY, allowing the job to be pending with the reason
"ReqNodeNotAvail, Reserved for maintenance".

    if (resv_overlap && bit_ffs(usable_node_mask) < 0) {

_build_node_list() proceeds to test the job but the node set bitmap always
has zero nodes because of the bit_and() with usable_node_mask:

    bit_and(node_set_ptr[node_set_inx].my_bitmap, usable_node_mask);

Which means that this if() condition is met on every iteration:

    if (node_set_ptr[node_set_inx].node_cnt == 0) {

_build_node_list() returns with ESLURM_REQUESTED_NODE_CONFIG_UNAVAILABLE.
This means the job is rejected on job submission.

By limiting the nodes available to the job in job_test_resv() to only those
matching the current partition being tested, in this situation
_build_node_list() will meet the if condition to allow the job to pend with
the reason "ReqNodeNotAvail, Reserved for maintenance".

Changelog: Fix handling of jobs submitted to a current reservation with
 flags OVERLAP,FLEX or OVERLAP,ANY_NODES when it overlaps nodes with a
 future maintenance reservation. When a job submission had a time limit that
 overlapped with the future maintenance reservation, it was rejected. Now
 the job is accepted but stays pending with the reason "ReqNodeNotAvail,
 Reserved for maintenance".
Ticket: 21831