jobcomp/kafka - Avoid unnecessary kafka message payload string copies
Performance improvement.

Instead of instructing librdkafka to make a copy of the payload and xfree()
the original string after producing, avoid such unnecessary copy and let
librdkafka hold the payload pointer to the originally allocated memory.

When the plugin regains ownership of the payload pointer again in the
delivery report callback then it'll finally be xfree()'d if needed (mainly
if the message is not produced again or queued for state save). This
also avoids unnecessary copies in the callback before queuing for state
save.

Basically no need for neither RD_KAFKA_MSG_F_COPY nor RD_KAFKA_MSG_F_FREE
message flags. The latter would use free() instead of xfree() anyway and
that would not be correct.

Issue: 50120
Ticket: 18189