slurmd - Do not catch SIGTTIN
This signal is delivered to a background process if it tries to read from
the terminal. The default behavior for that signal is to stop the program.

Ignoring this signal is not desired as slurmd in the background trying
to read from the console must be treated as an error. Moreover it was not
documented as being explicitly ignored by us.

This was introduced in b5fe169b079 and this commit reverts this change.

Ticket: 21828