slurmdbd - Do not catch SIGTTIN
This signal is delivered to a background process if it tries to read from
the terminal. The default behavior for that signal is to stop the program.

Ignoring this signal is not desired as slurmdbd in the background trying
to read from the console must be treated. Moreover it was not documented
as being explicitly ignored by us.

This was introduced in 5c584131e8 and this commit reverts this change.

Ticket: 21828