slurmdbd - Ignore SIGALRM signal as before the conversion to conmgr
Before commit 74b6c5f5a97 SIGALRM was caught and slurmdbd printed an error
informing of an invalid signal received, so the signal was effectively
ignored. After the commit, the signal is not explicitly ignored, so we
should catch it and at least print a debug message as with the other
ignored signals. Moreover it is documented that SIGALRM is ignored and we
were not ignoring it.

Ticket: 21828