slurmstepd - Ignore SIGALRM signal as before the conversion to conmgr
Before commit d5d2ba16b80 SIGALRM was masked and slurmstepd ignored it.
After the commit, the signal is not explicitly ignored, so we should
catch it and ignore it. Moreover it is documented that SIGALRM is ignored
and we were not honoring the docs.

Ticket: 21828
Cherry-picked: c0e2100dd4