Fix issue where gres with upper case name improperly registered
The slurmctld currently respects the configured case of gres, while the
slurmdbd sets all gres names to lower case. This has been the behavior
for as long as I can tell. We are not changing that here (though we do
recommend users just use lower case gres name to avoid any issues like
this one).

This commit fixes an issue that leads to Cluster Registered Tres in the
event table to be incorrectly set to 0 for gres with capitalized
characters in their names. This is from when the slurmctld doesn't
recognize the now lower case gres names from slurmdbd.

Changelog: Fix issue where gres with upper case names are improperly registered
Ticket: 15897