Docs - Elaborate on sshd AuthenticationMethods in pam_slurm_adopt page.
The current pam_slurm_adopt module is incompatible with the
keyboard-interactive authmethod. Failure to observe this warning
leads to SSH sessions not being properly adopted into the external
step (and thus killed once the job allocation is revoked).

The current documentation contains only a very oblique reference
to this behaviour, under the Limitations section.

Add a paragraph explicitly instructing administrators to ensure
keyboard-interactive is disabled, and describe the symptoms encountered
if this instruction is not observed.

Ticket: 21618
Cherry-picked: 2d6bb638