Command line MailType arguments will override batch script arguments
Previously the command line --mail-type argument would add to the batch
script arguments rather than override them. In the case of the NONE
argument, no change would be made since NONE would cause 0 to be 'OR'ed
into the MailType variable rather then the MailType variable being set to
0.

The command line arguments now override the batch script and behaves
as all other arguments. Besides that we corrected the behavior by matching
what is in the documentation and allowing comma-separated values for
different kinds of mail-type, and parsing the last mail-type argument and
not considering others. Before they were or'ed.

Changelog: MailType arguments set in the command line will override MailType
 arguments set in the batch script, and multiple --mail-type arguments will
 not be merged, honoring the documentation which states that multiple kinds
 of --mail-type can be set by providing a comma-separated list and making
 this behave like any other command line arguments.
Ticket: 21923