slurmscriptd - Catch and ignore the same signals ignored by slurmctld
Continuation of commits c883105f5b8 and 74b6c5f5a9.

A process signal mask is preserved after fork(). Before the conversion to
conmgr for signal handling in commit 74b6c5f5a9, slurmscriptd inherited the
signal mask from slurmctld. After that commit, slurmscriptd no longer
ignored some signals.

This was particularly problematic for SIGTSTP. If slurmctld runs in the
foreground and receives a SIGTSTP, then slurmctld catches and ignores that
signal. However, because slurmscriptd is in the same process group the shell
also sends SIGTSTP to slurmscriptd, which does not catch or mask SIGTSTP and
therefore is suspended. This results in a non-functional slurmscriptd, and
therefore a largely non-functional slurmctld that cannot even shut down
because it waits forever for slurmscriptd to respond that it is shutting
down. This is fixed by catching and ignoring SIGTSTP.

Commit c883105f5b8 attempted to revert behavior back to before commit
74b6c5f5a9. This commit completes that work.

Ticket: 21828
Cherry-picked: 85d98ea69b