mpi/pmix - Inject PMIX_APP_SIZE to PMIx KVS
Set the application size used for multiple program executions.

According to the PMIx standard v5.0 this is the number of processes in the
specified application, regardless of their execution state - i.e., this number
may include processes that either failed to start or have already terminated.

For Slurm this is just the number of tasks in this heterogeneous component.

Issue: 50072