Change slurmdb_tres_list_from_string() signature
Add list_t *sub_tres_list as a parameter so that an alternative to the
assoc_mgr_tres_list can be used. If a substitute tres list is provided the
assoc_mgr_locks are not grabbed. Passing NULL for the sub_tres_list will
result in the previous behavior.

Issue: 50233