burst_buffer/lua - Improve the state reason description
The state reason description was set in multiple places. Create a new
function to do this instead of repeating code. Improve the description,
making it clear if the job is held. Previously, many times the job would be
held, but it was not clear that the job was held from the state reason.

Signed-off-by: Michael Norris <michael.norris@schedmd.com>
Ticket: 21911