slurm_lua_loadscript() - Fix error handling for luaL_newstate()
Regression from commit ccff92007b2. Prior to that commit, if luaL_newstate()
failed the previous lua_State pointer was returned if it existed, or NULL if
not. Commit ccff92007b2 changed this to always return SLURM_SUCCESS if
luaL_newstate() failed.

Change the error handling to match the rest of this function: if a previous
lua_State exists, return success. Otherwise, return error.

Signed-off-by: Michael Norris <michael.norris@schedmd.com>
Ticket: 21911