Unify error handling code in slurm_lua_loadscript()
Combine the duplicate error handling code. In every error case, if a
previous lua_State exists, returns success; otherwise it returns error.
This makes it easy to move this error handling to the end.

The error messages are the same except for "use previous script". In one
case, "use previous script" was in the middle of the string, but in all
other cases it was at the end of the string. Unifying this code means it is
now at the end in all cases.

lua_pop() was sometimes used before lua_close(). This was unnecessary, since
lua_close() cleans up the lua_State anyway. This commit removes the
lua_pop() calls.

This commit sets the output err_msg parameter, if it is non-NULL.

Signed-off-by: Michael Norris <michael.norris@schedmd.com>
Ticket: 21911