_on_primary_msg - return SLURM_SUCCESS when rate limit exceeded
In case of rate_limit_exceeded() we need to send
SLURMCTLD_COMMUNICATIONS_BACKOFF to the client, but we shouldn't return
an error back to wrap_on_data. The logging is handled internally by
rate_limit_exceeded(), (with implementation of rl_log_freq=) and it's
not an error from slurmctld perspective.

Instead set rc to the value returned by slurm_send_rc_msg(), so we treat
as error the case when we can't send the message back to the client.

Ticket: 22009
Changelog: Prevent redundant "on_data returned rc: Rate limit exceeded,
 please retry momentarily" error message from being printed in
 slurmctld logs.
Cherry-picked: 6bd3f9a819