_validate_one_reservation - Don't remove reservation without valid users
Don't remove reservation even if there is no user with resolvable UID
now. Instead keep the list of UID's as it was. This may happen when site
experiences outage of IAM service, in that case we show an error in logs
but we don't remove the reservation. When external issues are resolved
reservation should work as before.

Ticket: 21504