Fix multi-partition, running job getting wrong partition on restart
For multi-partition requested jobs, the job's allocated partition is
state saved by putting the running partition's name at the beginning of
job_ptr->partition. On state load, the partition named list is converted
to a list of part_ptrs and it is expected that the running part_ptr is
the first part_ptr in the list. However, 88d0c160 changed it so that the
part_ptr_list is always sorted by PriorityTier, even on state load. So
the running job's partition was possibly being set to a different
partition than the one it was allocated.

This commit fixes it so that the job's part_ptr is set to the first
part_ptr in the list before sorting. This returns the running partition
for running jobs on state load. When the job is not running, it doesn't
matter what the part_ptr is set to because it will be changed throughout
the scheduling process of multiple partitions.

Ticket: 22010
Changelog: Fix running jobs, that requested multiple partitions, from
 potentially being set to the wrong partition on restart.
Cherry-picked: 0a2ed5b618